Add ?raw to script URL for dev (Vite 6 compatibility)
Vite 6 uses the Sec-Fetch-Dest header field to detect when a file should
be treated as a JavaScript module rather than a raw file. So when a
classic JS file is served in dev, it is served as a module. However,
we actually want the raw file in this situation, and we were relying on
the previous (Vite 5 and lower) behavior that did not detect this as a
JavaScript file.

The ?raw query ensures that the raw file is served (this is supported in
Vite 4 and 5 as well).

Change in Vite 6: https://github.com/vitejs/vite/pull/9981

Fixes: https://github.com/jack126guy/vite-plugin-classic-js/issues/2